fix(style-system): resolve medium-priority design system issues`n`n- PrimaryInput: add hover border state using border.hover token`n- WsTable: reduce header vertical padding s16->s8 to distinguish from data rows`n- InfoBar: change actions slot from FlexFit.tight to FlexFit.loose`n- MasterListCard: fix title text style from title(20px) to body2Bold(14px)

+++ packages/worklog_studio_style_system/lib/ui_kit/src/table/ws_table.dart
@@ -95,7 +95,7 @@ class WsTable<T> extends StatelessWidget {
       ),
       padding: EdgeInsets.symmetric(
         horizontal: theme.spacings.s16,
-        vertical: theme.spacings.s16,
+        vertical: theme.spacings.s8,
       ),
       child: Row(
         children: columns.asMap().entries.map((entry) {
